Use another heat calculation.
[Far Generation]Changed biome color data formating.
[Far Core]Changed world data provider temperature getting, which
required absolute temperature return.
[Far Core]Changed heat calculation.

+++ src/main/java/farcore/energy/thermal/IThermalHandlerBox.java
@@ -23,5 +23,5 @@ public interface IThermalHandlerBox extends IWorldSubContainer
 	 * @param value
 	 * @return Return true to prevent vanilla heat check.
 	 */
-	boolean onHeatChange(BlockPos source, BlockPos target, Direction sourceDir, double value);
+	boolean onHeatChange(BlockPos source, BlockPos target, Direction sourceDir, long value);
 }
